Remove deprecated makeStyles(); Move useFDA to api
