+++ src/pages/index.tsx
@@ -8,6 +8,7 @@ import SideMenu from "@/components/SubMenu";
 
 import StartStudyButton from "@/components/StartStudyButton";
 import AssignBatchNumberButton from "@/components/AssignBatchNumberButton";
+import LoginPage from "@/components/loginPage";
 
 export default function Home() {
   const [currentUser, setCurrentuser] = React.useState<Users>(Users.JHDoctor);
